search-os/requirements.txt
# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
fastparquet>=2023.10.0
openpyxl>=3.1.0  # Para leer archivos Excel

# AgGrid for large data visualization
streamlit-aggrid>=0.3.4

# AI/ML Integration
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
google-generativeai>=0.3.0

# External APIs
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Performance & Optimization
numpy>=1.26.0

# PDF Processing (temporal para análisis)
PyPDF2>=3.0.0
